add background-geolocation plugin and conditionally enable
